+++ CHANGELOG.md
@@ -1,3 +1,6 @@
 ## 0.1.0 (Unreleased)
 
 FEATURES:
+- cloudru-community provider
+- DNS Server Resource added
+- VPC Resource added